refactor(frontend): move simpleswitch to common
