+++ packages/code-du-travail-frontend/src/common/SimpleSwitch.js
